requirements.txt
selenium==4.32.0
pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
webdriver-manager==4.0.2
pyautogui==0.9.54 